Airflow/dags/requirements.txt
wheel
apache-airflow==2.1.2
apache-airflow-providers-postgres==2.0.0
confluent-kafka[avro]==1.7.0
psycopg2-binary==2.9.1
elastic-apm==6.3.3
